move qaqc app postgres utils to dcpy
