+++ apps/qaqc/example.env
@@ -1,4 +1,4 @@
 AWS_S3_ENDPOINT=
 AWS_SECRET_ACCESS_KEY=
 AWS_ACCESS_KEY_ID=
-SQL_ENGINE_EDM_DATA=
+BUILD_ENGINE=
